+++ src/components/SettingsPanel.tsx
@@ -38,7 +38,6 @@ import {
   ShortcutHint,
   screenshotShortcutKeys,
 } from './settings'
-import { IndexExpansionsPanel } from './IndexExpansionsPanel'
 
 type SettingItem = {
   id: SettingId
@@ -293,8 +292,6 @@ export function SettingsPanel() {
           ) : null}
         </ul>
       </SettingsSection>
-
-      <IndexExpansionsPanel />
 
       <SettingsSection title="Support" part="support">
         <ul className="settings-list">